feat: Update redirect path to `/photos/analog` in AlbumsPage
This commit updates the redirect path in the `AlbumsPage` component from `/photos/digi` to `/photos/analog`. The change ensures that users are redirected to the correct page when accessing the albums.

chore: Update npm dependencies

This commit updates the following npm dependencies to their latest stable versions:
- `clipboard-copy` to version `4.0.1`
- `framer-motion` to version `11.1.7`

fix: Update import path for MdxLayout component

This commit updates the import path for the `MdxLayout` component in the `page.tsx` file under the `src/app/projects/[projectname]` directory. The change ensures that the correct component is imported and used in the project page.

chore: Update CldImage component className

This commit updates the `className` prop value for the `CldImage` component in the `CldImage.tsx` file. The change adds additional CSS classes for a hover effect and transition animation, improving the visual experience for users.

chore: Update AlbumPage component

This commit updates the `AlbumPage` component in the `src/app/photos/[albumtype]/page.tsx` file. It adds a commented out paragraph element with hover and font styling, which can be used for future enhancements.

chore: Remove MdxLayout component

This commit removes the `MdxLayout` component from the project. The component was no longer being used and was deleted to declutter the codebase.

chore: Update ArrowBack component className

This commit updates the `className` prop value for the `ArrowBack` component in the `ArrowBack.tsx` file. The change adds a hover effect and transition animation to the SVG element, improving the interactivity for users.

chore: Update LogoImage component className

This commit updates the `className` prop value for the `LogoImage` component in the `LogoImage.tsx` file. The change adds a hover effect and transition animation to the anchor element, enhancing the visual experience for users.

chore: Update RootLayout component

This commit updates the `RootLayout` component in the `layout.tsx` file. It modifies the `title` and `description` metadata values to reflect the portfolio project of Louis Eggermont.

chore: Update tailwind.config.ts

This commit updates the `tailwind.config.ts` file. It extends the `scale` configuration to include a new scale value of `102`, allowing for a slight scaling effect in certain elements.

chore: Update AlbumPage component error handling

This commit updates the error handling in the `AlbumPage` component in the `src/app/photos/[albumtype]/[albumname]/page.tsx` file. It adds a try-catch block to handle any errors that occur during the fetching of pictures for the album. If an error occurs, a user-friendly error message is displayed to the user.

Note: The recent user commits and recent repository commits are identical.